ui: refactor recents view
- split components into different files and directories.
